+++ analyzer/windows/modules/packages/archive.py
@@ -14,25 +14,15 @@ except ImportError:
 
 from pathlib import Path
 
-from lib.common.abstracts import Package
-from lib.common.common import check_file_extension
 from lib.common.exceptions import CuckooPackageError
 from lib.common.hashing import hash_file
-from lib.common.parse_pe import choose_dll_export, is_pe_image
 from lib.common.results import upload_to_host
-from lib.common.zip_utils import extract_archive, get_file_names, winrar_extractor
+from lib.common.zip_utils import ArchivePackage, extract_archive, get_file_names, winrar_extractor, EXE_REGEX
 
 log = logging.getLogger(__name__)
 
 
-EXE_REGEX = re.compile(
-    r"(\.exe|\.dll|\.scr|\.msi|\.bat|\.lnk|\.js|\.jse|\.vbs|\.vbe|\.wsf|\.ps1|\.db|\.cmd|\.dat|\.tmp|\.temp|\.doc|\.xls)$",
-    flags=re.IGNORECASE,
-)
-PE_INDICATORS = [b"MZ", b"This program cannot be run in DOS mode"]
-
-
-class Archive(Package):
+class Archive(ArchivePackage):
     """Archive analysis package."""
 
     def __init__(self, options={}, config=None):
@@ -56,66 +46,6 @@ class Archive(Package):
         ("ProgramFiles", "Microsoft Office", "Office*", "EXCEL.EXE"),
         ("ProgramFiles", "Microsoft Office*", "root", "Office*", "EXCEL.EXE"),
     ]
-
-    def execute_interesting_file(self, root: str, file_name: str, file_path: str):
-        log.debug('file_name: "%s"', file_name)
-        if file_name.lower().endswith((".lnk", ".bat", ".cmd")):
-            cmd_path = self.get_path("cmd.exe")
-            cmd_args = f'/c "cd ^"{root}^" && start /wait ^"^" ^"{file_path}^"'
-            return self.execute(cmd_path, cmd_args, file_path)
-        elif file_name.lower().endswith(".msi"):
-            msi_path = self.get_path("msiexec.exe")
-            msi_args = f'/I "{file_path}"'
-            return self.execute(msi_path, msi_args, file_path)
-        elif file_name.lower().endswith((".js", ".jse", ".vbs", ".vbe", ".wsf")):
-            cmd_path = self.get_path("cmd.exe")
-            wscript = self.get_path_app_in_path("wscript.exe")
-            cmd_args = f'/c "cd ^"{root}^" && {wscript} ^"{file_path}^"'
-            return self.execute(cmd_path, cmd_args, file_path)
-        elif file_name.lower().endswith((".dll", ".db", ".dat", ".tmp", ".temp")):
-            # We are seeing techniques where dll files are named with the .db/.dat/.tmp/.temp extensions
-            if not file_name.lower().endswith(".dll"):
-                with open(file_path, "rb") as f:
-                    if not any(PE_indicator in f.read() for PE_indicator in PE_INDICATORS):
-                        return
-            dll_export = choose_dll_export(file_path)
-            if dll_export == "DllRegisterServer":
-                rundll32 = self.get_path("regsvr32.exe")
-            else:
-                rundll32 = self.get_path_app_in_path("rundll32.exe")
-                function = self.options.get("function", "#1")
-            arguments = self.options.get("arguments")
-            dllloader = self.options.get("dllloader")
-            dll_args = f'"{file_path}",{function}'
-            if arguments:
-                dll_args += f" {arguments}"
-            if dllloader:
-                newname = os.path.join(os.path.dirname(rundll32), dllloader)
-                shutil.copy(rundll32, newname)
-                rundll32 = newname
-            return self.execute(rundll32, dll_args, file_path)
-        elif file_name.lower().endswith(".ps1"):
-            powershell = self.get_path_app_in_path("powershell.exe")
-            args = f'-NoProfile -ExecutionPolicy bypass -File "{file_path}"'
-            return self.execute(powershell, args, file_path)
-        elif file_name.lower().endswith(".doc"):
-            # Try getting winword or wordview as a backup
-            try:
-                word = self.get_path_glob("WINWORD.EXE")
-            except CuckooPackageError:
-                word = self.get_path_glob("WORDVIEW.EXE")
-            return self.execute(word, f'"{file_path}" /q', file_path)
-        elif file_name.lower().endswith(".xls"):
-            # Try getting excel
-            excel = self.get_path_glob("EXCEL.EXE")
-            return self.execute(excel, f'"{file_path}" /q', file_path)
-        elif is_pe_image(file_path):
-            file_path = check_file_extension(file_path, ".exe")
-            return self.execute(file_path, self.options.get("arguments"), file_path)
-        else:
-            cmd_path = self.get_path("cmd.exe")
-            cmd_args = f'/c "cd ^"{root}^" && start /wait ^"^" ^"{file_path}^"'
-            return self.execute(cmd_path, cmd_args, file_path)
 
     def start(self, path):
         # 7za and 7r is limited so better install it inside of the vm
