+++ analyzer/windows/modules/packages/zip_compound.py
@@ -10,11 +10,12 @@ from typing import Tuple
 from lib.common.abstracts import Package
 from lib.common.exceptions import CuckooPackageError
 from lib.core.compound import create_custom_folders, extract_json_data
+from lib.common.zip_utils import ArchivePackage
 
 log = logging.getLogger(__name__)
 
 
-class ZipCompound(Package):
+class ZipCompound(ArchivePackage):
     """Extended functionality from the zip package to process compound samples"""
 
     PATHS = [
